requirements.txt
# --- WarSOC Backend Core ---
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.2
python-dotenv>=1.0.1
cryptography>=43.0.0

# --- Database & Real-time Messaging ---
motor>=3.6.0
redis>=5.0.0
hiredis>=3.0.0

# --- Pulse Agent & Communication ---
requests>=2.31.0
pyyaml>=6.0.2
# pywin32>=306  # Windows Event Log access — install ONLY on Windows agents (not the server)

# --- Security & Auth Bridge ---
python-multipart>=0.0.9    # REQUIRED for OAuth2 login [cite: 2026-02-20]
pyjwt[crypto]>=2.8.0      # REQUIRED for Agent JWTs [cite: 2026-02-18]
passlib[bcrypt]>=1.7.4     # REQUIRED for secure password storage [cite: 2026-02-20]

# --- Environment & Settings ---
pydantic-settings>=2.4.0   # REQUIRED for .env loading [cite: 2026-02-20]

# --- Real-time & Networking ---
websockets>=13.0           # REQUIRED for Live Alert popups [cite: 2026-02-26]
httpx>=0.27.0              # Async networking for the Worker [cite: 2026-02-18]

# --- Rate Limiting ---
slowapi>=0.1.9

# --- Additional runtime dependencies ---
email-validator>=1.3.1